Create Intro
Deep learning has seen tremendous progress over the past few years. This is largely due to the emergence of deep learning frameworks such as PyTorch and TensorFlow, which have greatly simplified even the most sophisticated research.
With smartphones having become the devices we use the most, the next wave of innovation is going to center on how we can leverage these rapid advances in deep learning to enhance our smartphone experiences. AI-powered mobile applications will become (and already are) smart enough to understand us and help us perform tasks via visual perception, language understanding, and voice recognition, even when not connected to the internet.
To accelerate the deployment of AI models on mobile devices, Facebook has just released PyTorch Mobile, which enables developers to deploy any PyTorch model to both Android and iOS. With PyTorch having become the most used deep learning framework for research, the mobile version will unlock a vast ecosystem of research and development that will ultimately enhance mobile experiences.
In this post, I’ll show how to take a PyTorch model trained on ImageNet and use it to build an Android application that can perform on-device image classification—taking a picture of any object and telling what it is.
Deploying A PyTorch model to Android requires the steps below:
Convert your model to TorchScript format (Python)
Add PyTorch Mobile as a Gradle dependency (Java)
Load your saved model with PyTorch Mobile to perform predictions (Java)
Setup
PyTorch
Visit pytorch.org to install the latest version of PyTorch for your operating system. If you already have a previous version installed, please upgrade to version 1.3.0 (minimum).
Android Studio
Install a recent version of Android Studio.
Model Conversion
To use our PyTorch model on Android, we need to convert it into TorchScript format. Luckily, this is quite an easy process. Below, we’ve loaded a pre-trained MobileNetV2 model, converted it into TorchScript, and saved it for use in our app. We can use any CNN architecture here; however, MobileNetV2 is highly optimized for both high speed and high accuracy on mobile devices.

Run this code and it will save your converted model as mobilenet-v2.pt—you can name your saved model anyway you wish.
It takes a lot of tweaking and fine-tuning to move from V1 of a mobile-ready model to one that’s ready for production. Fritz AI Studio is designed to streamline processes across the project lifecycle.
Building Our Mobile AI Application
Now that we have our model ready in a deployable format, fire up Android Studio, and create a new project named PytorchAndroid . While you can use any name, please stick to this naming for this tutorial to ensure code compatibility with my examples.
Step 1 : Add PyTorch Mobile to your Android project
With your project created in Android Studio, open the app’s build.gradle file and add PyTorch Mobile and TorchVision Mobile, as shown below:
implementation ‘org.pytorch:pytorch_android:1.3.0’
implementation ‘org.pytorch:pytorch_android_torchvision:1.3.0’

Android Studio will prompt you to synchronize your project. Click Sync Now and the dependencies will be downloaded.
Step 2: Add the model to your assets folder
Create an asset folder for your app by right clicking on app and navigating to New -> Folder -> Assets Folder. Now, copy the mobilenet-v2.pt file into the assets folder.
Step 3: Add the ImageNet label
In your main app package, create a file named Constants.java and put the contents of the file linked below into it:
johnolafenwa/PytorchMobile
You can't perform that action at this time. You signed in with another tab or window. You signed out in another tab or…
github.com

Note that I’m not previewing this file here, as a mapping of 1000 indexed names is too long.
Step 4: Add a Classifier class
In your main app package, create a file named Classifier.java and put the following in it.

This class is the most important part of our project—everything else is standard android stuffs. Hence, we shall break it down a bit.
This class defines the module and float arrays for the mean and standard deviation pre-processing. The model is loaded in the constructor.
The preprocess function takes in an image bitmap, resizes it to the specified size, performs mean and std pre-processing, and returns a tensor that can be input into our model.
The argmax function simply takes in scores and returns the index with the maximum score.
The predict function takes in any image bitmap, processes it into a tensor, runs it through the model to obtain a prediction, finds the maximum class using argmax, and finally, using the constants we created in step 3, it obtains the corresponding class name and returns it.
Step 5: Add the Utils class
Create the file Utils.java file containing the code below. It retrieves the absolute path to an asset file. We shall need this to retrieve the path to our model, which we added to assets earlier.

Kickstart your mobile machine learning projects — subscribe to the Fritz AI Newsletter for looks at common mobile ML problems, helpful resources, and much more.
Step 6: Add the Main Activity
Having created all our helpers and processors, add a new Basic Activity. Let the contents of the MainActivity.java file be as follows:

The activity_main.xml file should contain the following:

And content_main.xml:

Our main activity above does one simple thing. On the click of the button, it launches an external camera app. Once an image is captured, a bitmap is obtained from the data returned, and the classifier is used to predict the class of the image. The results are then passed to another activity for display.
Step 7: Add the Result Activity
Create a new Basic Activity named Result and put the following in the Result.java file:

Add this in the activity_result.xml:

Add this in the content_result.xml:

And that’s it! That’s all the files we need.
Building the app
Now, having followed all the steps above, build and run the application on an actual Android phone.
Below are screenshots from my phone:



You should test the app with a variety of items and see the result.
Summary
In this post, we’ve successfully built an Android application that’s able to recognize 1000 categories of items using a model that was originally trained with PyTorch.
This is intended as a reference and a starting point for doing a lot of amazing things with AI on Android, powered by PyTorch Mobile.